refactoring redux modules folder architecture
