--- src/store/modules/countries/actions.ts
@@ -1,34 +1,0 @@
-/* eslint-disable no-unused-vars */
-import { action } from 'typesafe-actions';
-
-import { CountriesTypes, Country } from './types';
-
-/**
- * Actions whos handle the request for all countries
- */
-
-export const handleLoadRequest = () => action(CountriesTypes.LOAD_REQUEST);
-
-export const handleLoadSuccess = (countries: Country[]) => action(CountriesTypes.LOAD_SUCCESS, { countries });
-
-export const handleLoadFailure = () => action(CountriesTypes.LOAD_FAILURE);
-
-/**
- * Actions whos handle the request for countries based on country name
- */
-
-export const handleLoadSearchRequest = (countryName: string) => action(CountriesTypes.LOAD_SEARCH_REQUEST, { countryName });
-
-export const handleLoadSearchSuccess = (country: Country[]) => action(CountriesTypes.LOAD_SEARCH_SUCCESS, { country });
-
-export const handleLoadSearchFailure = (message: string) => action(CountriesTypes.LOAD_SEARCH_FAILURE, { message });
-
-/**
- * Actions whos handle the request for countries based on continent
- */
-
-export const handleLoadFilterRequest = (continent: string) => action(CountriesTypes.LOAD_FILTER_REQUEST, { continent });
-
-export const handleLoadFilterSuccess = (country: Country[]) => action(CountriesTypes.LOAD_FILTER_SUCCESS, { country });
-
-export const handleLoadFilterFailure = (message: string) => action(CountriesTypes.LOAD_FILTER_FAILURE, { message });
