+++ src/store/modules/countries/actions/__tests__/unit/reducer.spec.ts
@@ -3,11 +3,11 @@
 /* eslint-disable no-unused-vars */
 import { Reducer } from 'redux-testkit';
 
-import handleExpectedAction from '../../../../../utils/tests/handleExpectedAction';
-import payload from '../../../../../utils/tests/payloadMock';
+import handleExpectedAction from '../../../../../../utils/tests/handleExpectedAction';
+import payload from '../../../../../../utils/tests/payloadMock';
 
-import countries from '../../reducer';
-import { CountriesState, CountriesTypes } from '../../types';
+import countries from '../../../reducer';
+import { CountriesState, CountriesTypes } from '../../../types';
 
 describe('reducer unit tests', () => {
   it('should have a initial state', () => {
